WIP : Refactor to a more concise code style
https://trello.com/c/MEfK1d0a
Finished time and log system refactor/Finished macro decorator/Use minimal polymorphism to decouple system.

+++ source/engine/common/InnoType.h
@@ -22,8 +22,8 @@ struct timeData
 	unsigned day;
 	unsigned hour;
 	unsigned minute;
-	unsigned second;
-	unsigned millisecond;
+	long long second;
+	long long millisecond;
 };
 
 enum class visiblilityType { INVISIBLE, BILLBOARD, STATIC_MESH, SKYBOX, GLASSWARE, EMISSIVE };
